+++ src/app/_component/Banner.tsx
@@ -1,3 +1,5 @@
+'use server';
+
 import * as styles from './banner.css';
 import { ComponentProps } from 'react';
 
@@ -5,7 +7,7 @@ type BannerProps = {
   color: string;
 };
 
-export default function Banner({ children, color }: ComponentProps<any> & BannerProps) {
+export default async function Banner({ children, color }: ComponentProps<any> & BannerProps) {
   return (
     <div className={styles.container} style={{ backgroundColor: color }}>
       {children}
