+++ src/app/_component/Feature.tsx
@@ -1,6 +1,8 @@
+'use server';
+
 import * as styles from './feature.css';
 
-export default function Feature() {
+export default async function Feature() {
   return (
     <section className={styles.container}>
       <img className={styles.image} src="/images/feature_info.png" alt="feature_info" />
